requirements.txt
bcrypt==5.0.0
blinker==1.9.0
click==8.3.1
cmake==4.2.1
colorama==0.4.6
Deprecated==1.3.1
dlib @ file:///C:/Users/david/OneDrive/Desktop/Documentacion/proyectos/auth-practice/dlib-19.24.1-cp311-cp311-win_amd64.whl#sha256=6f1a5ee167975d7952b28e0ce4495f1d9a77644761cf5720fb66d7c6188ae496
face-recognition==1.3.0
face_recognition_models==0.3.0
Flask==3.1.2
Flask-Bcrypt==1.0.1
Flask-Limiter==4.1.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-SQLAlchemy==3.1.1
greenlet==3.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
limits==5.6.0
MarkupSafe==3.0.3
numpy==1.26.4
opencv-python==4.10.0.84
ordered-set==4.1.0
packaging==25.0
pillow==12.1.0
pyotp==2.9.0
python-dotenv==1.2.1
qrcode==8.2
SQLAlchemy==2.0.45
typing_extensions==4.15.0
Werkzeug==3.1.5
wrapt==2.0.1